Initial commit
There's a bug that I don't understand. If I add 'cat' to the trie, this
works as expected. If I add a second word, 'car' to the trie, it appears
that a new trie is built.

It finds the first 'c' when add_word is called, but when it looks for
'a', it isn't found in the child.